add test granularity
