+++ tests/fixtures.py
@@ -1,0 +1,48 @@
+from pathlib import Path
+
+from ward import fixture
+
+from sliver import SliverClient, SliverClientConfig
+from sliver.pb.clientpb.client_pb2 import ImplantC2, ImplantConfig, OutputFormat
+
+
+class TestConstants:
+    def __init__(self):
+        self.multiplay_job_name = "grpc"
+
+
+@fixture(scope="global")
+def constants() -> TestConstants:
+    return TestConstants()
+
+
+@fixture(scope="global")
+async def sliver_client() -> SliverClient:
+    CONFIG_PATH = Path("~/.sliver-client/configs/sliverpy.cfg").expanduser()
+    config = SliverClientConfig.parse_config_file(CONFIG_PATH)
+    client = SliverClient(config)
+    await client.connect()
+    return client
+
+
+@fixture(scope="global")
+async def implant_config() -> ImplantConfig:
+    return ImplantConfig(
+        IsBeacon=False,
+        Name="sliver-pytest-" + os.urandom(8).hex(),
+        GOARCH="amd64",
+        GOOS="linux",
+        Format=OutputFormat.EXECUTABLE,
+        ObfuscateSymbols=False,
+        C2=[ImplantC2(Priority=0, URL="http://localhost:80")],
+    )
+
+
+@fixture(scope="global")
+def sliverpy_random_name() -> str:
+    return "sliver-pytest-" + os.urandom(8).hex()
+
+
+@fixture(scope="global")
+def data_dir() -> Path:
+    return Path(__file__).parent / "data"
